+++ packages/nextjs/components/ProductPage/StyleOptions.tsx
@@ -1,14 +1,17 @@
+"use client";
+
 import * as React from "react";
 import { H6, Pill } from "shared-ui";
-import { Style } from "utils/groqTypes/ProductDetail";
+import { ProductDetailVariants } from "utils/groqTypes/ProductDetail";
 
 interface Props {
-  options: Style[];
-  selectedStyle?: string;
-  onChange: (slicing: string) => void;
+  variant?: ProductDetailVariants[number];
 }
 
-export const StyleOptions = ({ options, selectedStyle, onChange }: Props) => {
+export const StyleOptions = ({ variant }: Props) => {
+  const options = variant?.style;
+  const [selectedStyle, setSelectedStyle] = React.useState(variant?.style?.[0]?.name || "");
+
   return (
     <div className="flex items-center justify-between">
       <H6>Style</H6>
@@ -17,7 +20,7 @@ export const StyleOptions = ({ options, selectedStyle, onChange }: Props) => {
           <Pill
             key={option?._id}
             selected={selectedStyle === option?.name}
-            onClick={() => onChange(option?.name ?? "")}
+            onClick={() => setSelectedStyle(option?.name ?? "")}
           >
             {option?.name}
           </Pill>
